fix: 全项目测试修复 - 修复29个问题
后端修复 (7个):
- AdminExamPaperController: @SuppressWarnings移至方法级别(Java8兼容)&类型改为Long
- H5UserController: 添加LearningRecord import,移除完全限定名
- H5PaymentController: 修复log变量遮蔽Lombok @Slf4j logger,补充orderId/amount
- ExamPaperService: List<Integer>改为List<Long>统一类型
- ExamPaperServiceImpl: 修改参数类型为Long
- QuestionServiceImpl: 移除冗余deleted检查
- 清理未使用的import

管理后台修复 (9个):
- 所有7个列表页面: el-pagination @change改为@current-change+@size-change
- Layout.vue: 修复el-avatar图标用法
- ExamList.vue: nextTick导入位置修复,移除脆弱的setTimeout,修复确认/取消逻辑
- 5个页面: 修复ElMessageBox.confirm取消时误显示错误消息
- StudentDetail.vue: 移除未使用的key变量

H5学员端修复 (8个):
- TabbarLayout/Profile/CourseDetail: 修复Vant4不存在的图标名称
- API拦截器: 移除不必要的reload
- VideoPlayer: async调用添加catch处理
- PracticeResult/ExamResult: JSON.parse添加try-catch
- ExamQuestion: autoSubmit添加await
- OrderList/MyWrongQuestions: 模拟数据补充缺失字段
- Home.vue: 移除未使用的reactive导入

+++ backend/src/main/java/com/onlinetrain/service/ExamPaperService.java
@@ -10,5 +10,5 @@ public interface ExamPaperService extends IService<ExamPaper> {
     /**
      * 保存试卷关联的题目
      */
-    void savePaperQuestions(Long paperId, List<Integer> questionIds);
+    void savePaperQuestions(Long paperId, List<Long> questionIds);
 }
